requirements.txt
python-telegram-bot==21.6
httpx==0.27.2
pydantic==2.8.2
python-dotenv==1.0.1
tenacity==8.5.0
pytest==8.2.2
pytest-asyncio==0.23.8
